Add argument "no.iff" -> switch off iff in gompred
The modified version of gompred accepts an additional
command line parameter "no-iff". If this parameter is
set to "1", the condition probabilities for inspected
pipes for are calculated differently. No pipe and
age-specific iff values are calculated and the transition matrixes that are used in the Markov Chain
are based on the marginal survival curves instead of
on the non-marginal survival curves.